move dvc functions to new generic dvc module
