+++ src/dmd_era5/__init__.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 from importlib.metadata import version
 
 from dmd_era5.config_reader import config_reader
+from dmd_era5.dvc_tools import add_data_to_dvc
 from dmd_era5.logger import log_and_print, setup_logger
 from dmd_era5.utils import create_mock_era5, slice_era5_dataset, thin_era5_dataset
 
@@ -17,5 +18,6 @@ __all__ = [
     "thin_era5_dataset",
     "slice_era5_dataset",
     "create_mock_era5",
+    "add_data_to_dvc",
 ]
 __version__ = version(__name__)
